requirements.txt
# Core
pytorch>=1.13 # if you're using pip, use torch
torchvision
numpy
scikit-learn
matplotlib
pandas
opencv-python-headless # opencv normal <-> if you're using pip, use opencv-python
facenet-pytorch

# Explainability
shap # If shap gives warnings related to numba, just lmk -- there's a known fix
lime

# Visualization
umap-learn
seaborn
mahotas
scikit-image


# Testing
pytest
pytest-cov

# Utility
tqdm
joblib
streamlit

# We're using large file system therefore in order to push huge files we need to install the mehtod that allows for it
#NOTE: brew install git-lfs
#NOTE: git lfs install